backend/requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.2
django-filter==1.1.0
django-oauth-toolkit==1.0.0
djangorestframework==3.7.7
idna==2.6
Markdown==2.6.11
oauth2-provider==0.0
oauthlib==2.0.6
pytz==2018.3
requests==2.18.4
urllib3==1.22
